Taxi service. Task 6. Commit 2. 1.Renaming the names of classes (controllers) to be more systematicity/logic. 2.Used one style for naming URLs. Rewrote the webservlet mapping annotations in controllers and links in jsp pages. 3.Instead of the ridiculous sortById method in DisplayAllDriversController.java, I implemented a comparator in the Driver.java class with an implementation of the compareTo method. + In DriverDaoImpl.java replaced HashSet with TreeSet implementation. 4.Minor corrections. timestamp: 00:56 30.08.2023
