+++ src/main/java/com/taxi/controller/AddCarController.java
@@ -1,7 +1,9 @@
 package com.taxi.controller;
 
 import com.taxi.config.AppConfig;
+import com.taxi.model.Car;
 import com.taxi.model.Manufacturer;
+import com.taxi.services.CarServiceImpl;
 import com.taxi.services.ManufacturerServiceImpl;
 import java.io.IOException;
 import javax.servlet.ServletException;
@@ -11,27 +13,31 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 
-@WebServlet("/ManufacturerController")
-public class ManufacturerController extends HttpServlet {
+@WebServlet("/cars/add")
+public class AddCarController extends HttpServlet {
     private final AnnotationConfigApplicationContext applicationContext
             = new AnnotationConfigApplicationContext(AppConfig.class);
+    private final CarServiceImpl carService
+            = applicationContext.getBean(CarServiceImpl.class);
     private final ManufacturerServiceImpl manufacturerService
             = applicationContext.getBean(ManufacturerServiceImpl.class);
-    private Manufacturer manufacturer
-            = new Manufacturer();
+    private final Car car
+            = new Car();
+    private Manufacturer manufacturer;
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        getServletContext().getRequestDispatcher("/WEB-INF/views/createManufacturer.jsp").forward(req, resp);
+        getServletContext().getRequestDispatcher("/WEB-INF/views/createCar.jsp").forward(req, resp);
     }
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        String name = req.getParameter("name");
-        manufacturer.setName(name);
-        String country = req.getParameter("country");
-        manufacturer.setCountry(country);
-        manufacturer = manufacturerService.create(manufacturer);
+        String model = req.getParameter("model");
+        String manufacturerId = req.getParameter("manufacturerId");
+        manufacturer = manufacturerService.get(Long.parseLong(manufacturerId));
+        car.setManufacturer(manufacturer);
+        car.setModel(model);
+        carService.create(car);
         resp.sendRedirect(req.getContextPath() + "/");
     }
 }
