+++ src/main/java/com/taxi/controller/AddDriverToCarController.java
@@ -13,7 +13,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 
-@WebServlet("/AddDriverToCarController")
+@WebServlet("/cars/drivers/add")
 public class AddDriverToCarController extends HttpServlet {
     private final AnnotationConfigApplicationContext applicationContext
             = new AnnotationConfigApplicationContext(AppConfig.class);
@@ -28,7 +28,7 @@ public class AddDriverToCarController extends HttpServlet {
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        getServletContext().getRequestDispatcher("/WEB-INF/views/addDriverToCar.jsp").forward(req, resp);
+        req.getRequestDispatcher("/WEB-INF/views/addDriverToCar.jsp").forward(req, resp);
     }
 
     @Override
@@ -36,11 +36,7 @@ public class AddDriverToCarController extends HttpServlet {
         String driverId = req.getParameter("driverId");
         String carId = req.getParameter("carId");
         driver = driverService.get(Long.parseLong(driverId));
-        System.out.println(Long.parseLong(driverId));
         car = carService.get(Long.parseLong(carId));
-        System.out.println(Long.parseLong(carId));
-        System.out.println(driver);
-        System.out.println(car);
         carService.addDriverToCar(driver, car);
         resp.sendRedirect(req.getContextPath() + "/");
     }
