+++ src/main/java/com/taxi/controller/AddManufacturerController.java
@@ -1,6 +1,7 @@
 package com.taxi.controller;
 
 import com.taxi.config.AppConfig;
+import com.taxi.model.Manufacturer;
 import com.taxi.services.ManufacturerServiceImpl;
 import java.io.IOException;
 import javax.servlet.ServletException;
@@ -10,19 +11,27 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 
-@WebServlet("/ManufacturerDeleteController/delete")
-public class ManufacturerDeleteController extends HttpServlet {
+@WebServlet("/manufacturers/add")
+public class AddManufacturerController extends HttpServlet {
     private final AnnotationConfigApplicationContext applicationContext
             = new AnnotationConfigApplicationContext(AppConfig.class);
-
     private final ManufacturerServiceImpl manufacturerService
             = applicationContext.getBean(ManufacturerServiceImpl.class);
+    private Manufacturer manufacturer
+            = new Manufacturer();
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        String manufacturerIdParam = req.getParameter("manufacturerId");
-        Long id = Long.parseLong(manufacturerIdParam);
-        manufacturerService.delete(id);
-        req.getRequestDispatcher("/WEB-INF/views/displayAllManufacturers.jsp").forward(req, resp);
+        getServletContext().getRequestDispatcher("/WEB-INF/views/createManufacturer.jsp").forward(req, resp);
+    }
+
+    @Override
+    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        String name = req.getParameter("name");
+        manufacturer.setName(name);
+        String country = req.getParameter("country");
+        manufacturer.setCountry(country);
+        manufacturer = manufacturerService.create(manufacturer);
+        resp.sendRedirect(req.getContextPath() + "/");
     }
 }
