+++ src/main/java/com/taxi/controller/DeleteCarController.java
@@ -1,10 +1,8 @@
 package com.taxi.controller;
 
 import com.taxi.config.AppConfig;
-import com.taxi.model.Car;
 import com.taxi.services.CarServiceImpl;
 import java.io.IOException;
-import java.util.List;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -12,8 +10,9 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 
-@WebServlet("/cars")
-public class DisplayAllCarsController extends HttpServlet {
+@WebServlet("/cars/delete")
+public class DeleteCarController extends HttpServlet {
+
     private final AnnotationConfigApplicationContext applicationContext
             = new AnnotationConfigApplicationContext(AppConfig.class);
 
@@ -22,8 +21,9 @@ public class DisplayAllCarsController extends HttpServlet {
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        List<Car> allCars = carService.getAll();
-        req.setAttribute("cars", allCars);
+        String carIdParam = req.getParameter("carId");
+        Long id = Long.parseLong(carIdParam);
+        carService.delete(id);
         req.getRequestDispatcher("/WEB-INF/views/displayAllCars.jsp").forward(req, resp);
     }
 }
