+++ src/main/java/com/taxi/controller/DeleteManufacturerController.java
@@ -1,7 +1,7 @@
 package com.taxi.controller;
 
 import com.taxi.config.AppConfig;
-import com.taxi.services.CarServiceImpl;
+import com.taxi.services.ManufacturerServiceImpl;
 import java.io.IOException;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -10,20 +10,19 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 
-@WebServlet("/CarDeleteController/delete")
-public class CarDeleteController extends HttpServlet {
-
+@WebServlet("/manufacturers/delete")
+public class DeleteManufacturerController extends HttpServlet {
     private final AnnotationConfigApplicationContext applicationContext
             = new AnnotationConfigApplicationContext(AppConfig.class);
 
-    private final CarServiceImpl carService
-            = applicationContext.getBean(CarServiceImpl.class);
+    private final ManufacturerServiceImpl manufacturerService
+            = applicationContext.getBean(ManufacturerServiceImpl.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        String carIdParam = req.getParameter("carId");
-        Long id = Long.parseLong(carIdParam);
-        carService.delete(id);
-        req.getRequestDispatcher("/WEB-INF/views/displayAllCars.jsp").forward(req, resp);
+        String manufacturerIdParam = req.getParameter("manufacturerId");
+        Long id = Long.parseLong(manufacturerIdParam);
+        manufacturerService.delete(id);
+        req.getRequestDispatcher("/WEB-INF/views/displayAllManufacturers.jsp").forward(req, resp);
     }
 }
